backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
psycopg2-binary
alembic
pydantic[email]
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart
redis
hiredis
asyncpg
python-dotenv
pandas
openpyxl
prometheus-client
httpx
requests